more refactoring and renaming for clarity
